webapi/tct-filesystem-tizen-tests/filesystem/File_deleteDirectory_with_onerror.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Samsung Electronics Co., Ltd.

Licensed under the Apache License, Version 2.0 (the License);
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.


Authors:
        Beata Koziarek <b.koziarek@samsung.com>
        Mariusz Polasinski <m.polasinski@samsung.com>


-->

<html>

<head>
<title>File_deleteDirectory_with_onerror</title>
<script type="text/javascript" src="../resources/testharness.js"></script>
<script type="text/javascript" src="../resources/testharnessreport.js"></script>
<script type="text/javascript" src="support/filesystem_common.js"></script>
</head>

<body>
<div id="log"></div>
<script type="text/javascript">

//==== TEST: File_deleteDirectory_with_onerror
//==== LABEL Check if File::deleteDirectory() method calls onerror properly (when removing non-existing directory)
//==== SPEC: Tizen Web API:IO:Filesystem:File:deleteDirectory M
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/filesystem.html
//==== TEST_CRITERIA MERRCB

var resolveSuccess, resolveError, deleteDirectorySuccess, deleteDirectoryError, fsTestDirName,
    expected = "NotFoundError", t = async_test(document.title);

t.step(function () {
    fsTestDirName =  getDirName("filesystemNoExist");

    deleteDirectorySuccess = t.step_func(function () {
        assert_unreached("deleteDirectorySuccess: deleteDirectoryError should be invoked");
    });

    deleteDirectoryError = t.step_func(function (error) {
        assert_equals(error.name, expected, "expect throw an exception");
        t.done();
    });

    resolveSuccess = t.step_func(function (dir) {
        dir.deleteDirectory(fsTestDirName, true, deleteDirectorySuccess, deleteDirectoryError);
    });

    resolveError = t.step_func(function (error) {
        assert_unreached("resolve() error callback invoked: name: " + error.name + ", msg: " + error.message);
    });

    prepareForTesting(t, function () {
        tizen.filesystem.resolve("documents", resolveSuccess, resolveError, "rw");
    });

});

</script>
</body>
</html>
